working advice bot
