+++ tsconfig.json
@@ -30,7 +30,6 @@
         "utils",
         "styles",
         "queries",
-        "services",
         "lib",
         "hooks",
         "functions",
